requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
